+++ blocklets/ai-studio/api/src/index.ts
@@ -19,16 +19,18 @@ import { projectCronManager } from './libs/cron-jobs';
 import { Config, isDevelopment } from './libs/env';
 import { NoPermissionError, NotFoundError } from './libs/error';
 import logger from './libs/logger';
+import { importPackageJson } from './libs/package-json';
 import { resourceManager } from './libs/resource';
 import { ensurePromptsEditor } from './libs/security';
 import routes from './routes';
+import { getOpenEmbed } from './routes/open-embed';
 import { wss } from './routes/ws';
 
 export const app = express();
 
 dotenv.config();
 
-const { name, version } = require('../../package.json');
+const { name, version } = importPackageJson();
 
 async function ensureUploadDirExists() {
   try {
@@ -48,6 +50,8 @@ app.use(cookieParser());
 app.use(express.json({ limit: '1 mb' }));
 app.use(express.urlencoded({ extended: true, limit: '1 mb' }));
 app.use(cors());
+
+app.get('/.well-known/blocklet/openembed', getOpenEmbed);
 
 app.use('/api', routes);