+++ blocklets/ai-studio/src/locales/en.ts
@@ -1,6 +1,6 @@
-import flat from 'flat';
+import { flatten } from 'flat';
 
-export default flat({
+export default flatten({
   name: 'Name',
   add: 'Add',
   project: 'Project',
@@ -773,4 +773,6 @@ export default flat({
     closed: 'Offline',
     closing: 'Closing',
   },
+  string: 'String',
+  arrayItem: 'Array Item',
 });