+++ blocklets/ai-studio/src/open-embed/agent-call.ts
@@ -1,0 +1,15 @@
+import { RunAgentInput, runAgent } from '@blocklet/aigne-sdk/api/agent';
+
+export default async function call({ input, aid }: { input: Omit<RunAgentInput, 'aid' | 'sessionId'>; aid: string }) {
+  const sessionId = Date.now().toString();
+
+  return runAgent({
+    inputs: {
+      question: input,
+    },
+    working: true,
+    sessionId,
+    aid,
+    responseType: 'text-stream',
+  });
+}