+++ packages/aigne-sdk/src/api/agent.ts
@@ -1,4 +1,4 @@
-import { RunAssistantResponse } from '@blocklet/ai-runtime/types';
+import { AssistantResponseType, RunAssistantResponse } from '@blocklet/ai-runtime/types';
 import { ResourceType } from '@blocklet/ai-runtime/types/resource';
 import { Agent, AgentWithConfig } from '@blocklet/ai-runtime/types/runtime/agent';
 import { fetchEventSource } from '@microsoft/fetch-event-source';
@@ -53,11 +53,17 @@ export async function runAgent({
 export async function runAgent({
   ...input
 }: RunAgentInput & { responseType: 'stream' }): Promise<ReadableStream<RunAssistantResponse>>;
-export async function runAgent({ responseType, ...input }: RunAgentInput & { responseType?: 'stream' }) {
+export async function runAgent({
+  ...input
+}: RunAgentInput & { responseType: 'text-stream' }): Promise<ReadableStream<string>>;
+export async function runAgent({
+  responseType,
+  ...input
+}: RunAgentInput & { responseType?: 'stream' | 'text-stream' }) {
   const path = '/api/ai/call';
 
-  if (responseType === 'stream') {
-    return new ReadableStream<RunAssistantResponse>({
+  if (responseType === 'stream' || responseType === 'text-stream') {
+    const stream = new ReadableStream<RunAssistantResponse>({
       async start(controller) {
         await fetchEventSource(joinURL(getMountPoint(AIGNE_RUNTIME_COMPONENT_DID), path), {
           openWhenHidden: true,
@@ -90,6 +96,18 @@ export async function runAgent({ responseType, ...input }: RunAgentInput & { res
         });
       },
     });
+
+    if (responseType === 'stream') return stream;
+
+    return stream.pipeThrough(
+      new TransformStream<RunAssistantResponse, string>({
+        transform: (chunk, controller) => {
+          if (chunk.type === AssistantResponseType.CHUNK && chunk.delta.content) {
+            controller.enqueue(chunk.delta.content);
+          }
+        },
+      })
+    );
   }
 
   return aigneRuntimeApi.post(path, input).then((res) => res.data);