+++ packages/code-editor/locales/zh.ts
@@ -1,0 +1,12 @@
+import { flatten } from 'flat';
+
+export default flatten({
+  settings: '设置',
+  editor: '编辑器',
+  format: '文档格式化',
+  vim: 'Vim 模式',
+  prettier: '格式化',
+  vimEnable: '启用 VIM',
+  adjustHeight: '调整高度',
+  memoryHeight: '记忆高度',
+});