+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.29 (2024-8-19)
+
+- feat(open-embed): new agent-view and agent-call embed components
+
+## 0.4.28 (2024-8-19)
+
+- chore: support object type input parameter
+
+## 0.4.27 (2024-8-16)
+
+- feat: support resize and full screen the code editor
+
 ## 0.4.26 (2024-8-9)
 
 - fix: number field cant input decimal